+++ modules/basics/index.json
@@ -1,0 +1,6 @@
+{
+    "name": "Basics",
+    "id": "basics",
+    "guarded": false,
+    "description": "No description was specified for this module."
+}
