+++ modules/fun/index.json
@@ -1,0 +1,6 @@
+{
+    "name": "Fun",
+    "id": "fun",
+    "guarded": false,
+    "description": "No description was specified for this module."
+}
