+++ modules/nadekoconnector/index.json
@@ -1,0 +1,6 @@
+{
+    "name": "Nadekoconnector",
+    "id": "nadekoconnector",
+    "guarded": false,
+    "description": "No description was specified for this module."
+}
