+++ modules/nadekoconnector/nc.js
@@ -28,7 +28,7 @@ module.exports = class NcSetCommand extends global.utils.baseCommand {
         if (!ctx.nadekoConnector)
             await ctx.db.set("nadekoconnector", { enabled: false });
         await ctx.message.delete();
-        let tryNc = new global.utils.nadekoConnector(ctx.args.address, ctx.args.password); 
+        let tryNc = new global.utils.nadekoConnector(ctx.args.address, ctx.args.password);
         let botInfo = await tryNc.getBotInfo();
         if (typeof botInfo.error !== "undefined") {
             await ctx.message.channel.send("Unable to connect to the given NadekoConnector instance.");
