+++ modules/utils/index.json
@@ -1,0 +1,6 @@
+{
+    "name": "Utils",
+    "id": "utils",
+    "guarded": false,
+    "description": "No description was specified for this module."
+}
