+++ package.json
@@ -1,10 +1,11 @@
 {
 	"name": "sapphyr",
-	"version": "0.0.2",
+	"version": "0.0.3",
 	"description": "A community oriented Discord bot",
 	"main": "main.js",
 	"scripts": {
-		"test": "echo \"Error: no test specified\" && exit 1"
+		"test": "echo \"Error: no test specified\" && exit 1",
+		"prod": "node --no-deprecation main.js"
 	},
 	"author": "rjt",
 	"license": "MIT",
